Move test for ops
